+++ spec/workato/extension/string_spec.rb
@@ -60,6 +60,71 @@ module Workato::Extension
       end
     end
 
+    describe '#rsa_sha512' do
+      let(:input) { "test\n" }
+      let(:pem_key) do
+        <<~PEM
+          -----BEGIN RSA PRIVATE KEY-----
+          MIIEogIBAAKCAQEAnzyis1ZjfNB0bBgKFMSvvkTtwlvBsaJq7S5wA+kzeVOVpVWw
+          kWdVha4s38XM/pa/yr47av7+z3VTmvDRyAHcaT92whREFpLv9cj5lTeJSibyr/Mr
+          m/YtjCZVWgaOYIhwrXwKLqPr/11inWsAkfIytvHWTxZYEcXLgAXFuUuaS3uF9gEi
+          NQwzGTU1v0FqkqTBr4B8nW3HCN47XUu0t8Y0e+lf4s4OxQawWD79J9/5d3Ry0vbV
+          3Am1FtGJiJvOwRsIfVChDpYStTcHTCMqtvWbV6L11BWkpzGXSW4Hv43qa+GSYOD2
+          QU68Mb59oSk2OB+BtOLpJofmbGEGgvmwyCI9MwIDAQABAoIBACiARq2wkltjtcjs
+          kFvZ7w1JAORHbEufEO1Eu27zOIlqbgyAcAl7q+/1bip4Z/x1IVES84/yTaM8p0go
+          amMhvgry/mS8vNi1BN2SAZEnb/7xSxbflb70bX9RHLJqKnp5GZe2jexw+wyXlwaM
+          +bclUCrh9e1ltH7IvUrRrQnFJfh+is1fRon9Co9Li0GwoN0x0byrrngU8Ak3Y6D9
+          D8GjQA4Elm94ST3izJv8iCOLSDBmzsPsXfcCUZfmTfZ5DbUDMbMxRnSo3nQeoKGC
+          0Lj9FkWcfmLcpGlSXTO+Ww1L7EGq+PT3NtRae1FZPwjddQ1/4V905kyQFLamAA5Y
+          lSpE2wkCgYEAy1OPLQcZt4NQnQzPz2SBJqQN2P5u3vXl+zNVKP8w4eBv0vWuJJF+
+          hkGNnSxXQrTkvDOIUddSKOzHHgSg4nY6K02ecyT0PPm/UZvtRpWrnBjcEVtHEJNp
+          bU9pLD5iZ0J9sbzPU/LxPmuAP2Bs8JmTn6aFRspFrP7W0s1Nmk2jsm0CgYEAyH0X
+          +jpoqxj4efZfkUrg5GbSEhf+dZglf0tTOA5bVg8IYwtmNk/pniLG/zI7c+GlTc9B
+          BwfMr59EzBq/eFMI7+LgXaVUsM/sS4Ry+yeK6SJx/otIMWtDfqxsLD8CPMCRvecC
+          2Pip4uSgrl0MOebl9XKp57GoaUWRWRHqwV4Y6h8CgYAZhI4mh4qZtnhKjY4TKDjx
+          QYufXSdLAi9v3FxmvchDwOgn4L+PRVdMwDNms2bsL0m5uPn104EzM6w1vzz1zwKz
+          5pTpPI0OjgWN13Tq8+PKvm/4Ga2MjgOgPWQkslulO/oMcXbPwWC3hcRdr9tcQtn9
+          Imf9n2spL/6EDFId+Hp/7QKBgAqlWdiXsWckdE1Fn91/NGHsc8syKvjjk1onDcw0
+          NvVi5vcba9oGdElJX3e9mxqUKMrw7msJJv1MX8LWyMQC5L6YNYHDfbPF1q5L4i8j
+          8mRex97UVokJQRRA452V2vCO6S5ETgpnad36de3MUxHgCOX3qL382Qx9/THVmbma
+          3YfRAoGAUxL/Eu5yvMK8SAt/dJK6FedngcM3JEFNplmtLYVLWhkIlNRGDwkg3I5K
+          y18Ae9n7dHVueyslrb6weq7dTkYDi3iOYRW8HRkIQh06wEdbxt0shTzAJvvCQfrB
+          jg/3747WSsf/zBTcHihTRBdAv6OmdhV4/dD5YBfLAkLrd+mX7iE=
+          -----END RSA PRIVATE KEY-----
+        PEM
+      end
+      let(:expected_base64) do
+        <<~BASE64.gsub("\n", '')
+          H6TtlEp4VlrjdRNPMbqCQmPouwei6NejJbDsXZRe/gw8QIPEv7DsbXk0+yDJx64c
+          VFQAVY2XfvETEMrNfmGnZADveT+iXTttJAERndSKNXttJ0nnCBqYX0JCkrn5LvuY
+          va3rm4+ajzvS+lQhUdLZsDqjaWDAPl6Rs38zx0TZ3Z7DAnOxdm18z5kw6EnA6EXf
+          6JcRCDhp2yFqg9kou5C1hpYJrIfrqbOR9rwdTSA3Q6OFK6LGcqE7re2x1dQk3b+S
+          7Zexnqi1Lg5WH78uU7Tm0Y76p8eBeh+Hi5q06+QsdChkRrd6fQ+O96gUbH/jpvJW
+          fxxbfz4nGHvVT9dEfztWiw==
+        BASE64
+      end
+
+      it 'signs a string correctly' do
+        expect(input.rsa_sha512(pem_key).base64).to eq(expected_base64)
+      end
+
+      context 'when binary input' do
+        let(:input) { Workato::Types::Binary.new("test\n") }
+
+        it 'signs a binary correctly' do
+          expect(input.rsa_sha512(pem_key).base64).to eq(expected_base64)
+        end
+      end
+
+      context 'when invalid pem key' do
+        let(:pem_key) { 'test' }
+
+        it 'raises error' do
+          expect { input.rsa_sha512(pem_key) }.to raise_error(Workato::Connector::Sdk::ArgumentError)
+        end
+      end
+    end
+
     describe '#hmac_sha256' do
       subject(:hmac_sha256) { input.hmac_sha256(key).to_json }
 
@@ -204,6 +269,18 @@ module Workato::Extension
       subject(:formula) { 'Hello'.encode_sha256.to_json }
 
       it { is_expected.to eq('0x185f8db32271fe25f561a6fc938b2e264306ec304eda518007d1764826381969') }
+    end
+
+    describe '"Hello".encode_sha512' do
+      subject(:formula) { 'Hello'.encode_sha512.to_json }
+
+      it { is_expected.to eq('0x3615f80c9d293ed7402687f94b22d58e529b8cc7916f8fac7fddf7fbd5af4cf777d3d795a7a00a16bf7e7f3fb9561ee9baae480da9fe7a18769e71886b03f315') } # rubocop:disable Layout/LineLength
+    end
+
+    describe '"Hello".encode_sha512_256' do
+      subject(:formula) { 'Hello'.encode_sha512_256.to_json }
+
+      it { is_expected.to eq('0x37653735623138623838643263623862653935623035656336313165353465323436303430386132646366383538663934353638363434366339643037616163') } # rubocop:disable Layout/LineLength
     end
 
     describe '"Hello".md5_hexdigest' do